+++ courseProgram/src/main/java/com/example/courseprogram/controller/DailyActivityController.java
@@ -38,13 +38,13 @@ public class DailyActivityController {
     //删除某学生的所有信息
     @PostMapping("/delete")
     public DataResponse deleteByStudentId(@RequestBody DataRequest dataRequest){
-        return dailyActivityService.deleteByStudentId(JsonUtil.parse(dataRequest.get("id"), Integer.class));
+        return dailyActivityService.deleteByStudentId(JsonUtil.parse(dataRequest.get("id"), Long.class));
     }
 
     //查找某学生的信息
     @PostMapping("/findByStudent")
     public DataResponse findByStudentId(@RequestBody DataRequest dataRequest){
-        return dailyActivityService.findByStudentId(JsonUtil.parse(dataRequest.get("id"), Integer.class));
+        return dailyActivityService.findByStudentId(JsonUtil.parse(dataRequest.get("id"), Long.class));
     }
 
     //获取所有数据
