+++ courseProgram/src/main/java/com/example/courseprogram/controller/HonorInfoController.java
@@ -33,13 +33,13 @@ public class HonorInfoController {
     //删除某学生的所有信息
     @PostMapping("/delete")
     public DataResponse deleteByStudentId(@RequestBody DataRequest dataRequest){
-        return honorInfoService.deleteByStudentId(JsonUtil.parse(dataRequest.get("id"), Integer.class));
+        return honorInfoService.deleteByStudentId(JsonUtil.parse(dataRequest.get("id"), Long.class));
     }
 
     //查找某学生的信息
     @PostMapping("/findByStudent")
     public DataResponse findByStudentId(@RequestBody DataRequest dataRequest){
-        return honorInfoService.findByStudentId(JsonUtil.parse(dataRequest.get("id"), Integer.class));
+        return honorInfoService.findByStudentId(JsonUtil.parse(dataRequest.get("id"), Long.class));
     }
 
     //获取所有数据
