+++ courseProgram/src/main/java/com/example/courseprogram/service/DailyActivityService.java
@@ -27,14 +27,14 @@ public class DailyActivityService {
     }
 
     //根据学生id删除
-    public DataResponse deleteByStudentId(Integer id){
+    public DataResponse deleteByStudentId(Long id){
         if(id==null)return DataResponse.failure(401,"信息不完整！");
         dailyActivityRepository.deleteByStudent_StudentId(id);
         return DataResponse.ok();
     }
 
     //根据学生id查找
-    public DataResponse findByStudentId(Integer id){
+    public DataResponse findByStudentId(Long id){
         if(id==null)return DataResponse.failure(401,"信息不完整！");
         List<DailyActivity> listA=dailyActivityRepository.findDailyActivitiesByStudent_StudentId(id);
         if(listA==null)return DataResponse.failure(404,"未找到该同学的信息");
