+++ courseProgram/src/main/java/com/example/courseprogram/service/HonorInfoService.java
@@ -39,14 +39,14 @@ public class HonorInfoService {
     }
 
 //    根据学生id删除
-    public DataResponse deleteByStudentId(Integer id){
+    public DataResponse deleteByStudentId(Long id){
         if(id==null)return DataResponse.failure(401,"信息不完整！");
         honorInfoRepository.deleteByStudent_StudentId(id);
         return DataResponse.okM("删除成功！");
     }
 
 //    根据学生id查找
-    public DataResponse findByStudentId(Integer id){
+    public DataResponse findByStudentId(Long id){
         if(id==null)return DataResponse.failure(401,"信息不完整！");
         List<HonorInfo> listA=honorInfoRepository.findHonorInfosByStudent_StudentId(id);
         if(listA==null)return DataResponse.failure(404,"未找到该同学的信息");
