+++ courseProgram/src/main/java/com/example/courseprogram/service/SocietyMemberService.java
@@ -27,14 +27,14 @@ public class SocietyMemberService {
     }
 
     //根据学生id删除
-    public DataResponse deleteByStudentId(Integer id){
+    public DataResponse deleteByStudentId(Long id){
         if(id==null)return DataResponse.failure(401,"信息不完整！");
         societyMemberRepository.deleteByStudent_StudentId(id);
         return DataResponse.ok();
     }
 
     //根据学生id查找
-    public DataResponse findByStudentId(Integer id){
+    public DataResponse findByStudentId(Long id){
         if(id==null)return DataResponse.failure(401,"信息不完整！");
         List<SocietyMember> listA=societyMemberRepository.findSocietyMembersByStudent_StudentId(id);
         if(listA==null)return DataResponse.failure(404,"未找到该同学的信息");
